requirements.txt
discord_webhook==1.3.0
espn_api==0.33.0
Flask==3.0.0
flask_cors==4.0.0
python-dotenv==1.0.0
tabulate==0.9.0
